backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
anthropic==0.39.0
openai==1.51.0
google-genai==1.0.0
aiosqlite==0.20.0
python-multipart==0.0.12
sse-starlette==2.1.3
markdown==3.7
weasyprint==62.3
aiofiles==24.1.0
